+++ packages/api/src/business/Resolvers.ts
@@ -11,7 +11,7 @@ import { TraitResolver } from "../models/Trait/TraitResolver.js";
 import { SpeciesVariantResolver } from "../models/SpeciesVariant/SpeciesVariantResolver.js";
 import { TraitListEntryResolver } from "../models/TraitListEntry/TraitListEntryResolver.js";
 import { LoginResolver } from "./LoginResolver.js";
-
+import { RoleResolver } from "../models/Role/RoleResolver.js";
 /**
  * This map contains all resolvers by their classnames. If you create a new resolver,
  * it must be added here to have its queries and mutations in the gql schema.
@@ -27,6 +27,7 @@ export const ResolversMap = {
   EnumValueSettingResolver,
   IdentityResolver,
   InviteCodeResolver,
+  RoleResolver,
 };
 
 export const ResolversArray = Object.values(ResolversMap);