+++ packages/client/src/ui/InviteCodeList/InviteCodeList.tsx
@@ -40,51 +40,31 @@ export const InviteCodeList: FC<InviteCodeListProps> = () => {
 
   const snackbarQueue = useSnackbar();
 
+  const headerCellProps = {
+    item: true,
+    xs: community ? 2 : 3,
+    component: Typography,
+    variant: "body1",
+    color: "text.secondary",
+    p: 1,
+  };
   if (data && isInviteCodeList(data.inviteCodes)) {
     return (
       <Grid container>
-        <Grid
-          item
-          xs={3}
-          component={Typography}
-          variant="body1"
-          color="text.secondary"
-          p={1}
-        >
+        <Grid {...headerCellProps} xs={3}>
           Invite Code
         </Grid>
-        <Grid
-          item
-          xs={3}
-          component={Typography}
-          variant="body1"
-          color="text.secondary"
-          p={1}
-        >
+        <Grid {...headerCellProps} xs={3}>
           Creator
         </Grid>
-        <Grid
-          item
-          xs={3}
-          component={Typography}
-          variant="body1"
-          color="text.secondary"
-          p={1}
-        >
-          Times Used
-        </Grid>
-        <Grid
-          item
-          xs={3}
-          component={Typography}
-          variant="body1"
-          color="text.secondary"
-          p={1}
-        >
-          Total Uses
-        </Grid>
+        {community && <Grid {...headerCellProps}>Role</Grid>}
+        <Grid {...headerCellProps}>Times Used</Grid>
+        <Grid {...headerCellProps}>Total Uses</Grid>
         {data.inviteCodes.list.map((inviteCode) => (
-          <GridRow key={inviteCode.id} xs={[3, 3, 3, 3]}>
+          <GridRow
+            key={inviteCode.id}
+            xs={community ? [3, 3, 2, 2, 2] : [3, 3, 3, 3]}
+          >
             <Stack alignItems="center" direction="row" p={2}>
               <IconButton
                 color="primary"
@@ -116,6 +96,11 @@ export const InviteCodeList: FC<InviteCodeListProps> = () => {
             <Typography variant="body1" p={2}>
               {inviteCode.creator.displayName}
             </Typography>
+            {community && (
+              <Typography variant="body1" p={2}>
+                {inviteCode.role?.name ?? ""}
+              </Typography>
+            )}
             <Typography variant="body1" p={2}>
               {inviteCode.claimCount}
             </Typography>